Minor improvement of the sync_event machinery. (#275)
It seems not easy for a newcomer to understand how exactly this machinery
works. Hence, add comments and improve the code style to make it more
conventional. Also, be stricter when there's a protocol breach detected and
raise a FATAL error instead of just ignoring it.

A database developer who uses these functions should understand that this
message is sent to the subscriber with the entire transaction if it was
successful. The LR progress status (commit LSN and other fields) will be
updated on the COMMIT of this transaction. So, the best practice is to call
sync_event() in a separate transaction.

Also, they should know that when the wait call finishes successfully, it
doesn't mean that all the transactions that began before the sync_event() call
have already arrived - it only means that all transactions committed before
the sync_event() call have already arrived and been processed on
the subscriber.

(cherry picked from commit 27e5857d35471828d254ae77c6f7436cc5089d9e)